Identify NTLMSSP response properly
DCERPC response is binary, so the comparison between the "substr"
of the raw response and hex string was failing.
Additionally the string used to check the result had the first
byte in last position.

Convert response segment to hex, and compare to the correct
response reference hex string.

Now failing during signing in DCERPC Authentication class with
NoMethodError undefined method `signing_key='