Initial commit of HTTP RC4 payload
Use reverse_http as base with recv and rev_http split into
different methods. Concatenate rev_http with recv_http_rc4 so as
to reuse the original connect-back code. Add mihi's suggestions
for preparing stack for the decryptor stub, and insert the actual
decryptor stub prior to returning into the payload.

Changed the generate_stage method to skip prepending the XORkey.

Compiles in metasm, fails at runtime. Analysis of the stack shows
that ECX looks to be short of the actual payload size by 4 when
the decryption routine is called.