Pass user-configured name to generated service PEs
The exe_sub_method in the Util::EXE Swiss Army Chainsaw looks for
a :servicename key in the passed opts hash to set a name in the
resulting binary.

The opts hash generated in Exploit::EXE's :exe_init_options method
doesn't include this configuration option, so the user is unable
to easily implement this substitution from the console/web UIs.

Add an advanced OptString to Exploit::EXE for 'EXE::ServiceName'
which will be processed during :generate_payload_exe_service after
:exe_init_options since its not relevant for normal PEs. The string
is passed as the :servicename option into the opts hash, to be
processed by Util::EXE's sub approach.