Reverse TCP x64 RC4 via max3raza's rc4_x64 asm
Max3raza graciously agreed to update the rc4 assembly block to x64.
This is the initial commit importing that asm work and creating a
new payload - x64/reverse_tcp_rc4.
As with x86, split out the reverse_tcp asm block into socket and
recv sections. Include the base payload in the new one, replacing
the recv block with a version modified from the x86 code to set
up the stack & registers for the rc4 decryptor.

Testing:
  None yet

Notes:
  My asm isn't great, and availability for non-prod work is even
less right now. Please excuse the cat-eating-bsod this is likely
to induce.